requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
google-generativeai
rank-bm25
pymupdf